refactor(bijou): extract textarea editor into textarea-editor.ts
Move the ~192-line interactive state machine from textarea.ts into a
dedicated textarea-editor.ts module:

- textarea-editor.ts: owns TextareaOptions interface and the
  interactiveTextarea() function with cursor navigation, scrolling,
  line numbers, placeholder, and max-length guard
- textarea.ts: slim facade with textarea() entry point, fallbackTextarea(),
  and a type re-export of TextareaOptions

Update ansi-lint.test.ts allowlist:
- Add textarea-editor.ts (inherited ANSI escapes)
- Remove textarea.ts (no longer contains ANSI escapes)
- Replace dag.ts with dag-render.ts (ANSI strip moved in prior commit)

Zero test modifications beyond allowlist. Zero index.ts modifications.
