+++ packages/bijou/src/core/ansi-lint.test.ts
@@ -27,12 +27,12 @@ const ALLOWED = new Set([
   'core/forms/select.ts',           // \x1b[K line-clear + key matching
   'core/forms/multiselect.ts',      // \x1b[K line-clear + key matching
   'core/forms/filter.ts',           // \x1b[K line-clear + key matching
-  'core/forms/textarea.ts',         // \x1b[K line-clear + key/arrow matching
+  'core/forms/textarea-editor.ts',  // \x1b[K line-clear + key/arrow matching
   'core/components/progress.ts',    // cursor control
   'core/components/spinner.ts',     // cursor control
   'core/components/hyperlink.ts',   // OSC 8 sequences
   'core/components/table.ts',       // ANSI strip regex
-  'core/components/dag.ts',         // ANSI strip regex
+  'core/components/dag-render.ts',  // ANSI strip regex
   'core/text/clip.ts',              // ANSI-aware text clipping
   'core/text/grapheme.ts',          // ANSI strip for width measurement
   'adapters/test/assertions.ts',    // assertion helpers reference ANSI patterns
